+++ cli/src/tui/mod.rs
@@ -9,7 +9,9 @@ mod component_page;
 mod dispatcher;
 mod event_editor;
 mod event_store;
+mod event_todo_editor;
 mod todo_editor;
 mod todo_store;
 
-pub use app::{draft_event, draft_todo, patch_event, patch_todo};
+pub use app::{draft_event, draft_event_or_todo, draft_todo, patch_event, patch_todo};
+pub use event_todo_editor::EventOrTodoDraft;